requirements.txt
discord.py
python-dotenv
requests
flask
psycopg2-binary